LayoutTests/imported/w3c/web-platform-tests/trusted-types/block-string-assignment-to-DOMWindowTimers-setTimeout-setInterval-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'want'
CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'want'
CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'want'

PASS window.setTimeout assigned via policy (successful Script transformation).
FAIL `window.setTimeout(string)` throws. assert_throws_js: function "_ => {
      setTimeout("window.timeoutTestString();");
    }" did not throw
FAIL `window.setTimeout(null)` throws. assert_throws_js: function "_ => {
      setTimeout(null);
    }" did not throw
PASS window.setInterval assigned via policy (successful Script transformation).
FAIL `window.setInterval(string)` throws. assert_throws_js: function "_ => {
      setInterval("window.intervalTestString()");
    }" did not throw
FAIL `window.setInterval(null)` throws. assert_throws_js: function "_ => {
      setInterval(null);
    }" did not throw
PASS `setTimeout(string)`, `setInterval(string)` via default policy (successful Script transformation).
PASS `setTimeout(null)`, `setInterval(null)` via default policy (successful Script transformation).
